integrates\refac(cross): #5749.7 estimated mttr
- Refactored the minTimeToRemediate attr type from float
to int as it is not meant to represent hours but minutes
in the front and back

+++ integrates/back/src/db_model/findings/types.py
@@ -154,7 +154,7 @@ class Finding(NamedTuple):
     approval: Optional[FindingState] = None
     creation: Optional[FindingState] = None
     description: str = ""
-    min_time_to_remediate: Optional[Decimal] = None
+    min_time_to_remediate: Optional[int] = None
     evidences: FindingEvidences = FindingEvidences()
     severity: Union[Finding20Severity, Finding31Severity] = Finding31Severity()
     sorts: FindingSorts = FindingSorts.NO
